+++ src/components/dropdown/dropdown.styles.ts
@@ -8,17 +8,28 @@ export default css`
     display: inline-block;
   }
 
-  .dropdown {
-    position: relative;
+  .dropdown::part(popup) {
+    z-index: var(--sl-z-index-dropdown);
+  }
+
+  .dropdown[data-current-placement^='top']::part(popup) {
+    transform-origin: bottom;
+  }
+
+  .dropdown[data-current-placement^='bottom']::part(popup) {
+    transform-origin: top;
+  }
+
+  .dropdown[data-current-placement^='left']::part(popup) {
+    transform-origin: right;
+  }
+
+  .dropdown[data-current-placement^='right']::part(popup) {
+    transform-origin: left;
   }
 
   .dropdown__trigger {
     display: block;
-  }
-
-  .dropdown__positioner {
-    position: absolute;
-    z-index: var(--sl-z-index-dropdown);
   }
 
   .dropdown__panel {
@@ -34,21 +45,5 @@ export default css`
 
   .dropdown--open .dropdown__panel {
     pointer-events: all;
-  }
-
-  .dropdown__positioner[data-placement^='top'] .dropdown__panel {
-    transform-origin: bottom;
-  }
-
-  .dropdown__positioner[data-placement^='bottom'] .dropdown__panel {
-    transform-origin: top;
-  }
-
-  .dropdown__positioner[data-placement^='left'] .dropdown__panel {
-    transform-origin: right;
-  }
-
-  .dropdown__positioner[data-placement^='right'] .dropdown__panel {
-    transform-origin: left;
   }
 `;